refactor(core): move workspace sql under worktree
